update status endpoint
