+++ src/eduid/scimapi/routers/status.py
@@ -4,10 +4,11 @@ from typing import Mapping
 from fastapi import Response
 
 from eduid.common.fastapi.context_request import ContextRequest
+from eduid.common.fastapi.utils import get_cached_response, set_cached_response
 from eduid.scimapi.api_router import APIRouter
 from eduid.scimapi.context_request import ScimApiRoute
 from eduid.scimapi.models.status import StatusResponse
-from eduid.scimapi.routers.utils.status import check_mongo, check_neo4j, get_cached_response, set_cached_response
+from eduid.scimapi.routers.utils.status import check_mongo, check_neo4j
 
 __author__ = "lundberg"
